test(ManageLegTraversal): Extract TraceData test class
